+++ src/components/RootComponent.tsx
@@ -4,7 +4,7 @@ import { Home } from './Home';
 import { Login } from './Login';
 import classes from './RootComponent.module.css';
 import { SignUp } from './SignUp';
-import { TweetCreate } from './Timeline/TweetCreate/TweetCreate';
+import { TweetCreateContainer } from './Timeline/TweetCreate/TweetCreateContainer';
 
 type Props = {
   isLogin: boolean;
@@ -38,7 +38,7 @@ export const RootComponent: React.FC<Props> = (props) => {
             exact
             path="/tweet"
             render={() => (
-              <TweetCreate
+              <TweetCreateContainer
                 isLogin={isLogin}
                 userImageURL="https://lh3.googleusercontent.com/6UgEjh8Xuts4nwdWzTnWH8QtLuHqRMUB7dp24JYVE2xcYzq4HA8hFfcAbU-R-PC_9uA1=w288-h288-n-rw"
                 userId={userId}
