+++ src/components/Timeline/TweetCreate/TweetCreateHeader.tsx
@@ -1,21 +1,18 @@
 import * as React from 'react';
 import ArrowBackIcon from '@material-ui/icons/ArrowBack';
-import { useHistory } from 'react-router-dom';
 import classes from './TweetCreateHeader.module.css';
 
-export const TweetCreateHeader: React.FC<{}> = () => {
-  const history = useHistory();
-  const Back = () => {
-    history.goBack();
-  };
+type Props = {
+  goBack(): void;
+  submitTweet(): void;
+};
 
-  const submitTweet = () => {
-    console.log('tweet button pushed.');
-  };
+export const TweetCreateHeader: React.FC<Props> = (props) => {
+  const { submitTweet, goBack } = props;
 
   return (
     <div className={classes.Header}>
-      <div className={classes.Icon} onClick={() => Back()}>
+      <div className={classes.Icon} onClick={() => goBack()}>
         <ArrowBackIcon />
       </div>
       <div className={classes.ButtonWrapper}>
